pyproject.toml
[project]
name = "multi-hop-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "google-genai>=1.33.0",
    "langchain>=0.3.27",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
]
